# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentState": {
        "additionalProperties": true,
        "properties": {
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Capabilities",
            "type": "object"
          },
          "clients": {
            "items": {
              "$ref": "#/components/schemas/ClientSummary"
            },
            "title": "Clients",
            "type": "array"
          },
          "config_revision": {
            "title": "Config Revision",
            "type": "integer"
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counters",
            "type": "object"
          },
          "discovery_port": {
            "title": "Discovery Port",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RuntimeEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Features",
            "type": "object"
          },
          "hints": {
            "items": {
              "type": "string"
            },
            "title": "Hints",
            "type": "array"
          },
          "listen": {
            "title": "Listen",
            "type": "string"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "paired_devices": {
            "items": {
              "$ref": "#/components/schemas/PairedDevice"
            },
            "title": "Paired Devices",
            "type": "array"
          },
          "pairing": {
            "$ref": "#/components/schemas/PairingState"
          },
          "pause_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Remaining"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Snapshot",
            "type": "object"
          },
          "token_set": {
            "title": "Token Set",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "config_revision",
          "platform",
          "listen",
          "hints",
          "token_set",
          "pairing",
          "discovery_port",
          "clients",
          "paired_devices",
          "capabilities",
          "features",
          "notifications",
          "snapshot",
          "logs",
          "events",
          "counters",
          "paused",
          "pause_remaining"
        ],
        "title": "AgentState",
        "type": "object"
      },
      "Apps": {
        "additionalProperties": true,
        "properties": {
          "apps": {
            "items": {
              "type": "string"
            },
            "title": "Apps",
            "type": "array"
          }
        },
        "required": [
          "apps"
        ],
        "title": "Apps",
        "type": "object"
      },
...
      "ClientSummary": {
        "additionalProperties": true,
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "device_id": {
            "default": "",
            "title": "Device Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "address"
        ],
        "title": "ClientSummary",
        "type": "object"
      },
...
      "DeviceRevoked": {
        "additionalProperties": true,
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "DeviceRevoked",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "NotificationStatus": {
        "additionalProperties": true,
        "properties": {
          "access": {
            "default": "unknown",
            "title": "Access",
            "type": "string"
          },
          "available": {
            "default": false,
            "title": "Available",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "error": {
            "default": "",
            "title": "Error",
            "type": "string"
          },
          "settings_action": {
            "default": "",
            "title": "Settings Action",
            "type": "string"
          }
        },
        "title": "NotificationStatus",
        "type": "object"
      },
      "ObsInput": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "timeout": {
            "title": "Timeout",
            "type": "number"
          }
        },
        "title": "ObsInput",
        "type": "object"
      },
...
      "ObsStatus": {
        "additionalProperties": true,
        "properties": {
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "current_scene": {
            "default": "",
            "title": "Current Scene",
            "type": "string"
          },
          "obs_version": {
            "default": "",
            "title": "Obs Version",
            "type": "string"
          },
          "scenes": {
            "items": {
              "type": "string"
            },
            "title": "Scenes",
            "type": "array"
          }
        },
        "required": [
          "connected"
        ],
        "title": "ObsStatus",
        "type": "object"
      },
...
      "PairingState": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          }
        },
        "required": [
          "required",
          "code",
          "expires_in"
        ],
        "title": "PairingState",
        "type": "object"
      },
      "PathRequest": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "file",
              "folder"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "PathRequest",
        "type": "object"
      },
      "PathSelection": {
        "additionalProperties": true,
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "file",
              "folder"
            ],
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "cancelled",
          "path",
          "kind"
        ],
        "title": "PathSelection",
        "type": "object"
      },
      "PermissionOpened": {
        "additionalProperties": true,
        "properties": {
          "opened": {
            "title": "Opened",
            "type": "boolean"
          },
          "permission": {
            "title": "Permission",
            "type": "string"
          }
        },
        "required": [
          "opened",
          "permission"
        ],
        "title": "PermissionOpened",
        "type": "object"
      },
      "PermissionRequest": {
        "additionalProperties": false,
        "properties": {
          "permission": {
            "title": "Permission",
            "type": "string"
          }
        },
        "required": [
          "permission"
        ],
        "title": "PermissionRequest",
        "type": "object"
      },
...
      "RuntimeEvent": {
        "additionalProperties": true,
        "properties": {
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Fields",
            "type": "object"
          },
          "level": {
            "enum": [
              "debug",
              "info",
              "warning",
              "error"
            ],
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "name",
          "level",
          "message",
          "fields"
        ],
        "title": "RuntimeEvent",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "LocalSession": {
        "description": "Per-start local UI session; never the durable console pairing token.",
        "in": "header",
        "name": "X-Deck3DS-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Authenticated loopback API. Console TCP and extension stdio protocols remain independent.",
    "title": "3Decks local agent API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/apps": {
      "get": {
        "operationId": "get_apps",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Apps"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Apps",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/obs/test": {
      "post": {
        "operationId": "test_obs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObsStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Obs",
        "tags": [
          "system"
        ]
      }
    },
    "/api/paired-devices/{device_id}": {
      "delete": {
        "operationId": "revoke_paired_device",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRevoked"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Revoke Device",
        "tags": [
          "system"
        ]
      }
    },
    "/api/pairing/rotate": {
      "post": {
        "operationId": "rotate_pairing",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Pairing",
        "tags": [
          "system"
        ]
      }
    },
    "/api/paths/pick": {
      "post": {
        "operationId": "pick_path",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathSelection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Pick",
        "tags": [
          "system"
        ]
      }
    },
    "/api/permissions/open": {
      "post": {
        "operationId": "open_permission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionOpened"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Permission",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/state": {
      "get": {
        "operationId": "get_state",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentState"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "State",
        "tags": [
          "system"
        ]
      }
    }
  },
  "security": [
    {
      "LocalSession": []
    }
  ]
}